misc: split lhr compat code to lib/lighthouse-compatibility.js
