+++ report/test/renderer/report-utils-test.js
@@ -94,176 +94,6 @@ describe('util helpers', () => {
   });
 
   describe('#prepareReportResult', () => {
-    describe('backward compatibility', () => {
-      it('corrects underscored `notApplicable` scoreDisplayMode', () => {
-        const clonedSampleResult = JSON.parse(JSON.stringify(sampleResult));
-
-        // Avoid injecting entity names into audits that would would
-        // make the diff at the end of this test difficult.
-        delete clonedSampleResult.entities;
-
-        let notApplicableCount = 0;
-        Object.values(clonedSampleResult.audits).forEach(audit => {
-          if (audit.scoreDisplayMode === 'notApplicable') {
-            notApplicableCount++;
-            audit.scoreDisplayMode = 'not_applicable';
-          }
-        });
-
-        assert.ok(notApplicableCount > 20); // Make sure something's being tested.
-
-        // Original audit results should be restored.
-        const preparedResult = ReportUtils.prepareReportResult(clonedSampleResult);
-
-        assert.deepStrictEqual(preparedResult.audits, sampleResult.audits);
-      });
-
-      it('corrects undefined auditDetails.type to `debugdata`', () => {
-        const clonedSampleResult = JSON.parse(JSON.stringify(sampleResult));
-
-        // Avoid injecting entity names into audits that would would
-        // make the diff at the end of this test difficult.
-        delete clonedSampleResult.entities;
-
-        // Delete debugdata details types.
-        let undefinedCount = 0;
-        for (const audit of Object.values(clonedSampleResult.audits)) {
-          if (audit.details && audit.details.type === 'debugdata') {
-            undefinedCount++;
-            delete audit.details.type;
-          }
-        }
-        assert.ok(undefinedCount > 4); // Make sure something's being tested.
-        assert.notDeepStrictEqual(clonedSampleResult.audits, sampleResult.audits);
-
-        // Original audit results should be restored.
-        const preparedResult = ReportUtils.prepareReportResult(clonedSampleResult);
-        assert.deepStrictEqual(preparedResult.audits, sampleResult.audits);
-      });
-
-      it('corrects `diagnostic` auditDetails.type to `debugdata`', () => {
-        const clonedSampleResult = JSON.parse(JSON.stringify(sampleResult));
-
-        // Avoid injecting entity names into audits that would would
-        // make the diff at the end of this test difficult.
-        delete clonedSampleResult.entities;
-
-        // Change debugdata details types.
-        let diagnosticCount = 0;
-        for (const audit of Object.values(clonedSampleResult.audits)) {
-          if (audit.details && audit.details.type === 'debugdata') {
-            diagnosticCount++;
-            audit.details.type = 'diagnostic';
-          }
-        }
-        assert.ok(diagnosticCount > 4); // Make sure something's being tested.
-        assert.notDeepStrictEqual(clonedSampleResult.audits, sampleResult.audits);
-
-        // Original audit results should be restored.
-        const preparedResult = ReportUtils.prepareReportResult(clonedSampleResult);
-        assert.deepStrictEqual(preparedResult.audits, sampleResult.audits);
-      });
-
-      it('corrects screenshots in the `filmstrip` auditDetails.type', () => {
-        const clonedSampleResult = JSON.parse(JSON.stringify(sampleResult));
-
-        // Avoid injecting entity names into audits that would would
-        // make the diff at the end of this test difficult.
-        delete clonedSampleResult.entities;
-
-        // Strip filmstrip screenshots of data URL prefix.
-        let filmstripCount = 0;
-        for (const audit of Object.values(clonedSampleResult.audits)) {
-          if (audit.details && audit.details.type === 'filmstrip') {
-            filmstripCount++;
-            for (const screenshot of audit.details.items) {
-              screenshot.data = screenshot.data.slice('data:image/jpeg;base64,'.length);
-            }
-          }
-        }
-        assert.ok(filmstripCount > 0); // Make sure something's being tested.
-        assert.notDeepStrictEqual(clonedSampleResult.audits, sampleResult.audits);
-
-        // Original audit results should be restored.
-        const preparedResult = ReportUtils.prepareReportResult(clonedSampleResult);
-        assert.deepStrictEqual(preparedResult.audits, sampleResult.audits);
-      });
-
-      it('moves full-page-screenshot audit', () => {
-        const clonedSampleResult = JSON.parse(JSON.stringify(sampleResult));
-
-        clonedSampleResult.audits['full-page-screenshot'] = {
-          details: {
-            type: 'full-page-screenshot',
-            ...sampleResult.fullPageScreenshot,
-          },
-        };
-        delete clonedSampleResult.fullPageScreenshot;
-        // Avoid injecting entity names into audits that would would
-        // make the diff at the end of this test difficult.
-        delete clonedSampleResult.entities;
-
-        assert.ok(clonedSampleResult.audits['full-page-screenshot'].details.nodes); // Make sure something's being tested.
-        assert.notDeepStrictEqual(clonedSampleResult.audits, sampleResult.audits);
-
-        // Original audit results should be restored.
-        const preparedResult = ReportUtils.prepareReportResult(clonedSampleResult);
-        assert.deepStrictEqual(preparedResult.audits, sampleResult.audits);
-        assert.deepStrictEqual(preparedResult.fullPageScreenshot, sampleResult.fullPageScreenshot);
-      });
-
-      it('corrects performance category without hidden group', () => {
-        const clonedSampleResult = JSON.parse(JSON.stringify(sampleResult));
-
-        clonedSampleResult.lighthouseVersion = '8.6.0';
-        delete clonedSampleResult.categoryGroups['hidden'];
-        for (const auditRef of clonedSampleResult.categories['performance'].auditRefs) {
-          if (auditRef.group === 'hidden') {
-            delete auditRef.group;
-          } else if (!auditRef.group) {
-            auditRef.group = 'diagnostics';
-          }
-        }
-        assert.notDeepStrictEqual(clonedSampleResult.categories, sampleResult.categories);
-        assert.notDeepStrictEqual(clonedSampleResult.categoryGroups, sampleResult.categoryGroups);
-
-        // Original audit results should be restored.
-        const clonedPreparedResult = ReportUtils.prepareReportResult(clonedSampleResult);
-        const preparedResult = ReportUtils.prepareReportResult(sampleResult);
-        assert.deepStrictEqual(clonedPreparedResult.categories, preparedResult.categories);
-        assert.deepStrictEqual(clonedPreparedResult.categoryGroups, preparedResult.categoryGroups);
-      });
-
-      it('converts old opportunity table column headings to consolidated table headings', () => {
-        const clonedSampleResult = JSON.parse(JSON.stringify(sampleResult));
-
-        // Avoid injecting entity names into audits that would would
-        // make the diff at the end of this test difficult.
-        delete clonedSampleResult.entities;
-
-        const auditsWithTableDetails = Object.values(clonedSampleResult.audits)
-          .filter(audit => audit.details?.type === 'table');
-        assert.notEqual(auditsWithTableDetails.length, 0);
-        for (const audit of auditsWithTableDetails) {
-          for (const heading of audit.details.headings) {
-            heading.itemType = heading.valueType;
-            heading.text = heading.label;
-            delete heading.valueType;
-            delete heading.label;
-
-            if (heading.subItemsHeading) {
-              heading.subItemsHeading.itemType = heading.subItemsHeading.valueType;
-              // @ts-expect-error
-              delete heading.subItemsHeading.valueType;
-            }
-          }
-        }
-
-        const preparedResult = ReportUtils.prepareReportResult(clonedSampleResult);
-        assert.deepStrictEqual(sampleResult.audits, preparedResult.audits);
-      });
-    });
-
     it('appends stack pack descriptions to auditRefs', () => {
       const clonedSampleResult = JSON.parse(JSON.stringify(sampleResult));
       const iconDataURL = 'data:image/svg+xml,%3Csvg xmlns="http://www.w3.org/2000/svg"%3E%3C/svg%3E';
