+++ report/tsconfig.json
@@ -14,6 +14,7 @@
   "include": [
     "**/*.js",
     "types/**/*.d.ts",
+    "../core/lib/lighthouse-compatibility.js",
   ],
   "exclude": [
     "generator/**/*.js",
