+++ tsconfig.json
@@ -77,6 +77,7 @@
     "core/test/lib/i18n/i18n-test.js",
     "core/test/lib/icons-test.js",
     "core/test/lib/lh-element-test.js",
+    "core/test/lib/lighthouse-compatibility-test.js",
     "core/test/lib/manifest-parser-test.js",
     "core/test/lib/median-run-test.js",
     "core/test/lib/minification-estimator-test.js",
